% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRelief.R
\name{addRelief}
\alias{addRelief}
\title{Add Map Bathy}
\usage{
addRelief(
  basemap,
  zlim = c(100, 3000),
  pal = paste0((grDevices::colorRampPalette(colors = c("lightgrey", "#000000")))(8),
    seq(10, 80, length.out = 8)),
  trim = T,
  refine = 0,
  verbose = T
)
}
\arguments{
\item{basemap}{a list object such as from plotBasemap()}

\item{zlim}{the value limits used to match up colors to numerical values. Default is range(z)}

\item{pal}{a vector of colors applied against the z grid to generate colors}

\item{trim}{a boolean value to turn on automatic trimming of the z grid to the visible map. Can cause (obvious) artefacts in some circumstances.}

\item{refine}{an integer value used to refine (or subsample) the provided grid. Positive values will refine and negative values will subsample.}

\item{verbose}{a boolean flag to turn on/off displayed messages}
}
\description{
Add a image layer to the map!
}
\author{
Thomas Bryce Kelly
}
