% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_1_3_4}
\alias{Bathy_1_3_4}
\title{Bathymetry data tile 1,3,4}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_1_3_4
}
\description{
Bathymetry data tile 1,3,4
}
\keyword{datasets}
