% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_2_10_4}
\alias{Bathy_2_10_4}
\title{Bathymetry data tile 2,10,4}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_2_10_4
}
\description{
Bathymetry data tile 2,10,4
}
\keyword{datasets}
