% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_2_12_2}
\alias{Bathy_2_12_2}
\title{Bathymetry data tile 2,12,2}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_2_12_2
}
\description{
Bathymetry data tile 2,12,2
}
\keyword{datasets}
