% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_2_1_1}
\alias{Bathy_2_1_1}
\title{Bathymetry data tile 2,1,1}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_2_1_1
}
\description{
Bathymetry data tile 2,1,1
}
\keyword{datasets}
