% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_2_2_0}
\alias{Bathy_2_2_0}
\title{Bathymetry data tile 2,2,0}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_2_2_0
}
\description{
Bathymetry data tile 2,2,0
}
\keyword{datasets}
