% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_3_0_2}
\alias{Bathy_3_0_2}
\title{Bathymetry data tile 3,0,2}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_3_0_2
}
\description{
Bathymetry data tile 3,0,2
}
\keyword{datasets}
