% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_3_4_3}
\alias{Bathy_3_4_3}
\title{Bathymetry data tile 3,4,3}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_3_4_3
}
\description{
Bathymetry data tile 3,4,3
}
\keyword{datasets}
