% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_4_0_3}
\alias{Bathy_4_0_3}
\title{Bathymetry data tile 4,0,3}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_4_0_3
}
\description{
Bathymetry data tile 4,0,3
}
\keyword{datasets}
