% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_4_7_2}
\alias{Bathy_4_7_2}
\title{Bathymetry data tile 4,7,2}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_4_7_2
}
\description{
Bathymetry data tile 4,7,2
}
\keyword{datasets}
