% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_4_8_4}
\alias{Bathy_4_8_4}
\title{Bathymetry data tile 4,8,4}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_4_8_4
}
\description{
Bathymetry data tile 4,8,4
}
\keyword{datasets}
