% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bathy_4_9_5}
\alias{Bathy_4_9_5}
\title{Bathymetry data tile 4,9,5}
\format{
An object of class \code{list} of length 3.
}
\usage{
Bathy_4_9_5
}
\description{
Bathymetry data tile 4,9,5
}
\keyword{datasets}
