% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findElevation.R
\name{findElevation}
\alias{findElevation}
\title{Retreive elevation data for a location}
\usage{
findElevation(lon, lat)
}
\arguments{
\item{lon}{Longitude values to get elevation for}

\item{lat}{Latitude values to get elevation for}
}
\description{
Retreive elevation data for a location
}
