% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTiles.R
\name{getTiles}
\alias{getTiles}
\title{Add bathymetry tiles}
\usage{
getTiles(lon, lat, n, verbose = F)
}
\arguments{
\item{lon}{a vector or array of longitude values that match the z grid provided}

\item{lat}{a vector or array of latitude values that match the z grid provided}

\item{n}{a grid (matrix or array) to be used for plotting}

\item{verbose}{a boolean flag to turn on/off displayed messages}
}
\description{
Add bathymetry tiles
}
\author{
Thomas Bryce Kelly
}
