% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileMap.R
\name{tileMapLon}
\alias{tileMapLon}
\title{Determine longitude index for tiling}
\usage{
tileMapLon(lon, delta = 30, offset = 180)
}
\arguments{
\item{lon}{a vector of longitude values}

\item{delta}{the width of each tile in degrees}

\item{offset}{offset}
}
\description{
Determine longitude index for tiling
}
\author{
Thomas Bryce Kelly
}
