% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileMap.R
\name{tileMapLat}
\alias{tileMapLat}
\title{Determine latitude index for tiling}
\usage{
tileMapLat(lat, delta = 30, offset = 90)
}
\arguments{
\item{lat}{a vector of latitude values}

\item{delta}{the height of each tile in degrees}

\item{offset}{offset}
}
\description{
Determine latitude index for tiling
}
\author{
Thomas Bryce Kelly
}
